services/web/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
gunicorn==20.0.4
slackclient==2.7.2
slackeventsapi==2.2.0
sqlalchemy==1.3.18
aiocron==1.3
